requirements.txt
streamlit
openai-whisper
torch
transformers
fpdf2
numpy
pandas
